refactor(Charts): deduplicate types across multiple components
