+++ libs/ui-react-visualization/src/lib/Components/CartesianChart/index.ts
@@ -1,3 +1,3 @@
 export { CartesianChart } from './CartesianChart';
 export { useCartesianChartContext } from './context';
-export type { CartesianChartProps } from './types';
+export type { CartesianChartContextValue, CartesianChartProps } from './types';
