+++ libs/ui-react-visualization/src/lib/utils/index.ts
@@ -1,5 +1,4 @@
 export type {
-  CartesianChartContextValue,
   CategoricalScale,
   ChartInset,
   ChartScaleFunction,
