+++ libs/ui-react-visualization/src/lib/utils/types.ts
@@ -1,5 +1,4 @@
 import type { ScaleBand, ScaleLinear, ScaleLogarithmic } from 'd3-scale';
-import type { BaseAxisProps } from '../Components/Axis';
 
 export type ChartInset = {
   top: number;
@@ -77,38 +76,3 @@ export type NumericScale =
 export type CategoricalScale = ScaleBand<number>;
 
 export type ChartScaleFunction = NumericScale | CategoricalScale;
-
-export type CartesianChartContextValue = {
-  /**
-   * All data series registered in the chart.
-   */
-  series: Series[];
-  /**
-   * Lookup map for series by ID. Stable reference when series identity is unchanged.
-   */
-  seriesMap: Map<string, Series>;
-  /**
-   * Returns the x-axis scale. Accepts an optional axis ID for future multi-axis support.
-   */
-  getXScale: (id?: string) => ChartScaleFunction | undefined;
-  /**
-   * Returns the y-axis scale. Accepts an optional axis ID for future multi-axis support.
-   */
-  getYScale: (id?: string) => ChartScaleFunction | undefined;
-  /**
-   * Returns the x-axis config. Accepts an optional axis ID for future multi-axis support.
-   */
-  getXAxisConfig: (id?: string) => BaseAxisProps | undefined;
-  /**
-   * Returns the y-axis config. Accepts an optional axis ID for future multi-axis support.
-   */
-  getYAxisConfig: (id?: string) => BaseAxisProps | undefined;
-  /**
-   * Pixel bounds of the drawable region.
-   */
-  drawingArea: DrawingArea;
-  /**
-   * Number of data points along the x-axis.
-   */
-  dataLength: number;
-};
